requirements.txt
pandas>=1.3.0,<1.4.0
jupyter>=1.0.0,<2.0.0
notebook>=6.4.0,<7.0.0
matplotlib>=3.5.0,<3.6.0
seaborn>=0.11.0,<0.12.0
plotly>=5.0.0,<6.0.0
nbconvert>=6.0.0,<7.0.0
numpy>=1.20.0,<1.22.0
scipy>=1.7.0,<1.8.0
gspread<5.0.0
colorama
gspread-dataframe
google-auth
lxml[html_clean]